requirements.txt
beautifulsoup4>=4.14.2
langdetect>=1.0.9
lxml>=6.0.2
marimo>=0.17.6
numpy>=1.24.3
opencv-python-headless>=4.8.1.78
pillow>=9.5.0
pydantic>=2.6.1
rfdetr>=1.3.0
shapely>=2.0.6
transformers>=4.29.0,<4.47.0
yattag>=1.16.1

fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9